requirements.txt
pandas
apache-airflow-providers-snowflake